% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_create.R
\name{boundary_create}
\alias{boundary_create}
\title{Create bounding box based on buffer size}
\usage{
boundary_create(points_shape, buffer_radius, pjctn_crs)
}
\arguments{
\item{points_shape}{points to be buffered (must be sf object)}

\item{buffer_radius}{distance in meters}

\item{pjctn_crs}{projection with meters unit}
}
\value{
returns a bounding box
}
\description{
Create bounding box based on buffer size
}
\examples{
\dontrun{
boundary_create(points_sf, 750, 102039)
}
}
