% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_read.R
\name{points_read}
\alias{points_read}
\title{Read user-supplied file of points}
\usage{
points_read(points_file)
}
\arguments{
\item{points_file}{path to an sf object file}
}
\value{
a dataframe
}
\description{
Read user-supplied file of points
}
\examples{
\dontrun{
points_read("path/to/my/file")
}
}
