% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roadstructure_append.R
\name{road_structure_append}
\alias{road_structure_append}
\title{Title}
\usage{
road_structure_append(roads_shp, structure_shp)
}
\arguments{
\item{roads_shp}{roads (sf object)}

\item{structure_shp}{structures (such as buildings and walls) (sf object)}
}
\value{
sf object containing all structures and roads
}
\description{
Title
}
